Added role to user model, fixed some http error codes, added login and register with jwt token

+++ server/routes/orders.js
@@ -29,27 +29,38 @@ router.post("/", async (req, res) => {
     const savedOrder = await order.save();
     res.status(201).json(savedOrder);
   } catch (err) {
-    res.status(404).json({ message: err });
+    res.status(500).json({ message: err });
   }
 });
 
 router.get("/:orderId", async (req, res) => {
   try {
     const order = await Order.findById(req.params.orderId);
+    if (!order)
+      return res
+        .status(404)
+        .json({ message: "Order with this ID doesn't exist!" });
+
     res.status(200).json(order);
   } catch (err) {
-    res.status(404).json({ message: err });
+    res.status(500).json({ message: err });
   }
 });
 
 router.delete("/:orderId", async (req, res) => {
   try {
+    const order = await Order.findById(req.params.orderId);
+    if (!order)
+      return res
+        .status(404)
+        .json({ message: "Order with this ID doesn't exist!" });
+
     const removedOrder = await Order.deleteOne({
       _id: req.params.orderId,
     });
     res.status(200).json(removedOrder);
   } catch (err) {
-    res.status(404).json({ message: err });
+    res.status(500).json({ message: err });
   }
 });
